Updated "We are humble"
I strongly disagree that we are exceptional people. I think what makes a company great is that it creates an environment where otherwise normal people can do exceptional work. We all have things that make us unique, but we're not all that exceptional. What is remarkable, is that together, we're able to create something that is actually quite exceptional.