Adding a value about recognizing value
Pricing as a discussion point was always going to be a difficult conversation at dbt Labs. It can feel actively antithetical to some values (Profits are exhaust, Values are more important than success) and can make us concerned about culture creep generally. 

Nevertheless, it is both an overcorrection and unfair to our front-of-house team members to treat pricing as, at best, the third rail. Given that, it might benefit us as a company to treat pricing as another part of being a business, in that we must define our relationship with it rather than defining ourselves without it. 

This value would seek to capture the goals we have around pricing, namely creating a structure that:
- appropriately values our effort, particularly in the face of our open source solution remaining available. 
- aligns with the resources required to do our best work and support it beyond release. 
- respects our customers, with price tiering reflective of comparable feature tiering.
- allows our sales and business development teams to empower the business without being functionally isolated from our values.